requirements.txt
python-dotenv==1.0.1
python_telegram_bot==21.4
requests==2.32.3
telegram==0.0.1
paramiko==3.4.0